Split parser::value
