+++ ruststep/src/parser/exchange/anchor.rs
@@ -1,4 +1,4 @@
-use crate::parser::{combinator::*, token::*};
+use crate::parser::{combinator::*, token::*, value::*};
 use nom::{branch::alt, Parser};
 
 /// anchor_section = `ANCHOR;` [anchor_list] `ENDSEC;` .
