+++ ruststep/src/parser/value.rs
@@ -1,0 +1,23 @@
+use serde::Deserialize;
+
+/// Left hand side value
+#[derive(Debug, Clone, PartialEq, Deserialize)]
+pub enum LValue {
+    /// Like `#11`
+    Entity(u64),
+    /// Like `@11`
+    Value(u64),
+}
+
+/// Right hand side value
+#[derive(Debug, Clone, PartialEq, Deserialize)]
+pub enum RValue {
+    /// Like `#11`
+    Entity(u64),
+    /// Like `@11`
+    Value(u64),
+    /// Like `#CONST_ENTITY`
+    ConstantEntity(String),
+    /// Like `@CONST_VALUE`
+    ConstantValue(String),
+}
