+++ ruststep/src/tables.rs
@@ -1,3 +1,4 @@
+use crate::parser::value::RValue;
 use serde::Deserialize;
 use std::{
     collections::HashMap,
